dart ffigen doesn't support uint64_t
but if we need to support a diversifier above  2,147,483,647, we will have to split the 64-bit BigInt diversifier in Dart into two 32-bit int32s.

+++ src/dart_interface.h
@@ -1,7 +1,7 @@
 #ifndef ORG_FIRO_SPARK_DART_INTERFACE_H
 #define ORG_FIRO_SPARK_DART_INTERFACE_H
 
-const char* getAddress(const char* keyData, int index, uint64_t diversifier);
+const char* getAddress(const char* keyData, int index, int diversifier);
 
 /*
 const char *createFullViewKey(const char* keyData, int index);
